refactor: move user-presenter to its own folder
